pyproject.toml
[tool.poetry]
name = "mpddq"
version = "0.1.0"
description = ""
authors = ["David Hoyes <dphoyes@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-mpd2 = "^3.0.4"
anyio = {git = "git@github.com:agronholm/anyio.git", tag = "3.0.0rc4"}
PyYAML = "^5.4.1"
asyncinotify = "^2.0.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
mpddq = 'mpddq:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
